avoid name conflicts by renaming sqlglot module
